consolidating buffer classes
