+++ Magnefu/Magnefu.vcxproj
@@ -149,7 +149,7 @@
     <ClInclude Include="src\CacheableResource.h" />
     <ClInclude Include="src\Camera.h" />
     <ClInclude Include="src\Globals.h" />
-    <ClInclude Include="src\IndexBuffer.h" />
+    <ClInclude Include="src\Magnefu\Renderer\Buffer.h" />
     <ClInclude Include="src\Light.h" />
     <ClInclude Include="src\Magnefu.h" />
     <ClInclude Include="src\Magnefu\Application.h" />
@@ -176,26 +176,20 @@
     <ClInclude Include="src\Maths\PrimitiveCommon.h" />
     <ClInclude Include="src\Maths\Quaternion.h" />
     <ClInclude Include="src\Maths\Vectors.h" />
-    <ClInclude Include="src\Mesh.h" />
-    <ClInclude Include="src\Model.h" />
+    <ClInclude Include="src\Platform\OpenGL\OpenGLBuffer.h" />
     <ClInclude Include="src\Platform\OpenGL\OpenGLContext.h" />
     <ClInclude Include="src\Platform\Windows\WindowsInput.h" />
     <ClInclude Include="src\Platform\Windows\WindowsWindow.h" />
-    <ClInclude Include="src\Renderer.h" />
+    <ClInclude Include="src\Magnefu\Renderer\Renderer.h" />
     <ClInclude Include="src\ResourceCache.h" />
     <ClInclude Include="src\Shader.h" />
     <ClInclude Include="src\Texture.h" />
     <ClInclude Include="src\Timer.h" />
-    <ClInclude Include="src\VertexArray.h" />
-    <ClInclude Include="src\VertexBuffer.h" />
-    <ClInclude Include="src\VertexBufferAttribsLayout.h" />
+    <ClInclude Include="src\Magnefu\Renderer\VertexArray.h" />
+    <ClInclude Include="src\Magnefu\Renderer\VertexBufferAttribsLayout.h" />
     <ClInclude Include="src\mfpch.h" />
     <ClInclude Include="src\tests\Test.h" />
     <ClInclude Include="src\tests\Test2DTexture.h" />
-    <ClInclude Include="src\tests\Test3DRender.h" />
-    <ClInclude Include="src\tests\TestBatchRendering.h" />
-    <ClInclude Include="src\tests\TestLighting.h" />
-    <ClInclude Include="src\tests\TestModelLoading.h" />
     <ClInclude Include="vendor\GLAD\include\KHR\khrplatform.h" />
     <ClInclude Include="vendor\GLAD\include\glad\glad.h" />
     <ClInclude Include="vendor\GLFW\deps\getopt.h" />
@@ -359,7 +353,7 @@
   <ItemGroup>
     <ClCompile Include="src\Camera.cpp" />
     <ClCompile Include="src\Globals.cpp" />
-    <ClCompile Include="src\IndexBuffer.cpp" />
+    <ClCompile Include="src\Magnefu\Renderer\Buffer.cpp" />
     <ClCompile Include="src\Light.cpp" />
     <ClCompile Include="src\Magnefu\Application.cpp" />
     <ClCompile Include="src\Magnefu\Core.cpp" />
@@ -374,28 +368,22 @@
     <ClCompile Include="src\Maths\Primitive3D.cpp" />
     <ClCompile Include="src\Maths\Quaternion.cpp" />
     <ClCompile Include="src\Maths\Vectors.cpp" />
-    <ClCompile Include="src\Mesh.cpp" />
-    <ClCompile Include="src\Model.cpp" />
+    <ClCompile Include="src\Platform\OpenGL\OpenGLBuffer.cpp" />
     <ClCompile Include="src\Platform\OpenGL\OpenGLContext.cpp" />
     <ClCompile Include="src\Platform\Windows\WindowsInput.cpp" />
     <ClCompile Include="src\Platform\Windows\WindowsWindow.cpp" />
-    <ClCompile Include="src\Renderer.cpp" />
+    <ClCompile Include="src\Magnefu\Renderer\Renderer.cpp" />
     <ClCompile Include="src\ResourceCache.cpp" />
     <ClCompile Include="src\Shader.cpp" />
     <ClCompile Include="src\Texture.cpp" />
     <ClCompile Include="src\Timer.cpp" />
-    <ClCompile Include="src\VertexArray.cpp" />
-    <ClCompile Include="src\VertexBuffer.cpp" />
-    <ClCompile Include="src\VertexBufferAttribsLayout.cpp" />
+    <ClCompile Include="src\Magnefu\Renderer\VertexArray.cpp" />
+    <ClCompile Include="src\Magnefu\Renderer\VertexBufferAttribsLayout.cpp" />
     <ClCompile Include="src\mfpch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="src\tests\Test.cpp" />
     <ClCompile Include="src\tests\Test2DTexture.cpp" />
-    <ClCompile Include="src\tests\Test3DRender.cpp" />
-    <ClCompile Include="src\tests\TestBatchRendering.cpp" />
-    <ClCompile Include="src\tests\TestLighting.cpp" />
-    <ClCompile Include="src\tests\TestModelLoading.cpp" />
     <ClCompile Include="vendor\glm\detail\glm.cpp" />
     <ClCompile Include="vendor\imgui\backends\imgui_impl_glfw.cpp" />
     <ClCompile Include="vendor\imgui\backends\imgui_impl_opengl3.cpp" />
