+++ Magnefu/Magnefu.vcxproj.filters
@@ -117,9 +117,6 @@
     <ClInclude Include="src\Globals.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="src\IndexBuffer.h">
-      <Filter>src</Filter>
-    </ClInclude>
     <ClInclude Include="src\Light.h">
       <Filter>src</Filter>
     </ClInclude>
@@ -198,12 +195,6 @@
     <ClInclude Include="src\Maths\Vectors.h">
       <Filter>src\Maths</Filter>
     </ClInclude>
-    <ClInclude Include="src\Mesh.h">
-      <Filter>src</Filter>
-    </ClInclude>
-    <ClInclude Include="src\Model.h">
-      <Filter>src</Filter>
-    </ClInclude>
     <ClInclude Include="src\Platform\OpenGL\OpenGLContext.h">
       <Filter>src\Platform\OpenGL</Filter>
     </ClInclude>
@@ -213,7 +204,7 @@
     <ClInclude Include="src\Platform\Windows\WindowsWindow.h">
       <Filter>src\Platform\Windows</Filter>
     </ClInclude>
-    <ClInclude Include="src\Renderer.h">
+    <ClInclude Include="src\Magnefu\Renderer\Renderer.h">
       <Filter>src</Filter>
     </ClInclude>
     <ClInclude Include="src\ResourceCache.h">
@@ -228,13 +219,10 @@
     <ClInclude Include="src\Timer.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="src\VertexArray.h">
+    <ClInclude Include="src\Magnefu\Renderer\VertexArray.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="src\VertexBuffer.h">
-      <Filter>src</Filter>
-    </ClInclude>
-    <ClInclude Include="src\VertexBufferAttribsLayout.h">
+    <ClInclude Include="src\Magnefu\Renderer\VertexBufferAttribsLayout.h">
       <Filter>src</Filter>
     </ClInclude>
     <ClInclude Include="src\mfpch.h">
@@ -244,18 +232,6 @@
       <Filter>src\tests</Filter>
     </ClInclude>
     <ClInclude Include="src\tests\Test2DTexture.h">
-      <Filter>src\tests</Filter>
-    </ClInclude>
-    <ClInclude Include="src\tests\Test3DRender.h">
-      <Filter>src\tests</Filter>
-    </ClInclude>
-    <ClInclude Include="src\tests\TestBatchRendering.h">
-      <Filter>src\tests</Filter>
-    </ClInclude>
-    <ClInclude Include="src\tests\TestLighting.h">
-      <Filter>src\tests</Filter>
-    </ClInclude>
-    <ClInclude Include="src\tests\TestModelLoading.h">
       <Filter>src\tests</Filter>
     </ClInclude>
     <ClInclude Include="vendor\GLAD\include\KHR\khrplatform.h">
@@ -735,15 +711,14 @@
     <ClInclude Include="vendor\stb_image\stb_image.h">
       <Filter>vendor\stb_image</Filter>
     </ClInclude>
+    <ClInclude Include="src\Magnefu\Renderer\Buffer.h" />
+    <ClInclude Include="src\Platform\OpenGL\OpenGLBuffer.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Camera.cpp">
       <Filter>src</Filter>
     </ClCompile>
     <ClCompile Include="src\Globals.cpp">
-      <Filter>src</Filter>
-    </ClCompile>
-    <ClCompile Include="src\IndexBuffer.cpp">
       <Filter>src</Filter>
     </ClCompile>
     <ClCompile Include="src\Light.cpp">
@@ -788,12 +763,6 @@
     <ClCompile Include="src\Maths\Vectors.cpp">
       <Filter>src\Maths</Filter>
     </ClCompile>
-    <ClCompile Include="src\Mesh.cpp">
-      <Filter>src</Filter>
-    </ClCompile>
-    <ClCompile Include="src\Model.cpp">
-      <Filter>src</Filter>
-    </ClCompile>
     <ClCompile Include="src\Platform\OpenGL\OpenGLContext.cpp">
       <Filter>src\Platform\OpenGL</Filter>
     </ClCompile>
@@ -803,7 +772,7 @@
     <ClCompile Include="src\Platform\Windows\WindowsWindow.cpp">
       <Filter>src\Platform\Windows</Filter>
     </ClCompile>
-    <ClCompile Include="src\Renderer.cpp">
+    <ClCompile Include="src\Magnefu\Renderer\Renderer.cpp">
       <Filter>src</Filter>
     </ClCompile>
     <ClCompile Include="src\ResourceCache.cpp">
@@ -818,13 +787,10 @@
     <ClCompile Include="src\Timer.cpp">
       <Filter>src</Filter>
     </ClCompile>
-    <ClCompile Include="src\VertexArray.cpp">
+    <ClCompile Include="src\Magnefu\Renderer\VertexArray.cpp">
       <Filter>src</Filter>
     </ClCompile>
-    <ClCompile Include="src\VertexBuffer.cpp">
-      <Filter>src</Filter>
-    </ClCompile>
-    <ClCompile Include="src\VertexBufferAttribsLayout.cpp">
+    <ClCompile Include="src\Magnefu\Renderer\VertexBufferAttribsLayout.cpp">
       <Filter>src</Filter>
     </ClCompile>
     <ClCompile Include="src\mfpch.cpp">
@@ -834,18 +800,6 @@
       <Filter>src\tests</Filter>
     </ClCompile>
     <ClCompile Include="src\tests\Test2DTexture.cpp">
-      <Filter>src\tests</Filter>
-    </ClCompile>
-    <ClCompile Include="src\tests\Test3DRender.cpp">
-      <Filter>src\tests</Filter>
-    </ClCompile>
-    <ClCompile Include="src\tests\TestBatchRendering.cpp">
-      <Filter>src\tests</Filter>
-    </ClCompile>
-    <ClCompile Include="src\tests\TestLighting.cpp">
-      <Filter>src\tests</Filter>
-    </ClCompile>
-    <ClCompile Include="src\tests\TestModelLoading.cpp">
       <Filter>src\tests</Filter>
     </ClCompile>
     <ClCompile Include="vendor\glm\detail\glm.cpp">
@@ -875,5 +829,7 @@
     <ClCompile Include="vendor\stb_image\stb_image.cpp">
       <Filter>vendor\stb_image</Filter>
     </ClCompile>
+    <ClCompile Include="src\Magnefu\Renderer\Buffer.cpp" />
+    <ClCompile Include="src\Platform\OpenGL\OpenGLBuffer.cpp" />
   </ItemGroup>
 </Project>
