--- Magnefu/src/IndexBuffer.cpp
@@ -1,44 +1,0 @@
-#include "mfpch.h"
-
-#include "IndexBuffer.h"
-
-#include <GLAD/glad.h>
-
-IndexBuffer::IndexBuffer(unsigned int count, const unsigned int* data)
-    : m_Count(count)
-{
-    ASSERT(sizeof(unsigned int) == sizeof(GLuint));
-
-    glGenBuffers(1, &m_RendererID);
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_RendererID);
-    if (data)
-    {
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(unsigned int), data, GL_STATIC_DRAW);
-    }
-    else
-    {
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(unsigned int), data, GL_DYNAMIC_DRAW);
-    }
-
-
-}
-
-IndexBuffer::~IndexBuffer()
-{
-    glDeleteBuffers(1, &m_RendererID);
-}
-
-void IndexBuffer::Bind() const
-{
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_RendererID);
-}
-
-void IndexBuffer::Unbind() const
-{
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
-}
-
-void IndexBuffer::InvalidateBuffer() const
-{
-    glInvalidateBufferData(m_RendererID);
-}
