+++ Magnefu/src/Magnefu/Renderer/Buffer.h
@@ -1,0 +1,32 @@
+#pragma once
+
+
+namespace Magnefu
+{
+	class VertexBuffer
+	{
+	public:
+		virtual ~VertexBuffer() = default;
+
+		virtual void Bind() const = 0;
+		virtual void Unbind() const = 0;
+
+		static VertexBuffer* Create(uint32_t size);
+		static VertexBuffer* Create(uint32_t size, float* data);
+	};
+
+	
+
+	class IndexBuffer
+	{
+	public:
+		virtual ~IndexBuffer() = default;
+
+		virtual void Bind() const = 0;
+		virtual void Unbind() const = 0;
+
+		virtual uint32_t GetCount() const = 0;
+
+		static IndexBuffer* Create(uint32_t count, uint32_t* data);
+	};
+}
