+++ Magnefu/src/Platform/OpenGL/OpenGLBuffer.cpp
@@ -1,0 +1,70 @@
+#include "mfpch.h"
+#include "OpenGLBuffer.h"
+
+#include "GLAD/glad.h"
+
+namespace Magnefu
+{
+	OpenGLVertexBuffer::OpenGLVertexBuffer(uint32_t size, float* data)
+	{
+		glCreateBuffers(1, &m_RendererID);
+		glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
+
+		if (data)
+		{
+			glBufferData(GL_ARRAY_BUFFER, size, data, GL_STATIC_DRAW);
+		}
+		else
+		{
+			glBufferData(GL_ARRAY_BUFFER, size, data, GL_DYNAMIC_DRAW);
+		}
+	}
+
+	OpenGLVertexBuffer::~OpenGLVertexBuffer()
+	{
+		glDeleteBuffers(1, &m_RendererID);
+	}
+
+	void OpenGLVertexBuffer::Bind() const
+	{
+		glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
+	}
+
+	void OpenGLVertexBuffer::Unbind() const
+	{
+		glBindBuffer(GL_ARRAY_BUFFER, 0);
+	}
+
+
+	//// INDEXBUFFER ////
+
+	OpenGLIndexBuffer::OpenGLIndexBuffer(uint32_t count, uint32_t* data) :
+		m_Count(count)
+	{
+		glCreateBuffers(1, &m_RendererID);
+		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_RendererID);
+		if (data)
+		{
+			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(uint32_t), data, GL_STATIC_DRAW);
+		}
+		else
+		{
+			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(uint32_t), data, GL_DYNAMIC_DRAW);
+		}
+	}
+
+	OpenGLIndexBuffer::~OpenGLIndexBuffer()
+	{
+		glDeleteBuffers(1, &m_RendererID);
+	}
+
+	void OpenGLIndexBuffer::Bind() const
+	{
+		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_RendererID);
+	}
+
+	void OpenGLIndexBuffer::Unbind() const
+	{
+		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
+	}
+}
