--- Magnefu/src/VertexArray.h
@@ -1,19 +1,0 @@
-#pragma once
-
-class VertexBuffer;
-class VertexBufferAttribsLayout;
-
-class VertexArray
-{
-private:
-	unsigned int m_RendererID;
-
-public:
-	VertexArray();
-	~VertexArray();
-
-	void Unbind() const;
-	void Bind()   const;
-
-	void AddBuffer(const VertexBuffer& vb, const VertexBufferAttribsLayout& layout );
-};
