--- Magnefu/src/VertexBuffer.cpp
@@ -1,35 +1,0 @@
-#include "mfpch.h"
-
-#include "VertexBuffer.h"
-#include <GLAD/glad.h>
-
-VertexBuffer::VertexBuffer(unsigned int size, const void* data)
-{
-    glGenBuffers(1, &m_RendererID);
-    glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
-
-    if (data)
-    {
-        glBufferData(GL_ARRAY_BUFFER, size, data, GL_STATIC_DRAW);
-    }
-    else
-    {
-        glBufferData(GL_ARRAY_BUFFER, size, data, GL_DYNAMIC_DRAW);
-    }
-        
-}
-
-VertexBuffer::~VertexBuffer()
-{
-    glDeleteBuffers(1, &m_RendererID);
-}
-
-void VertexBuffer::Bind() const
-{
-    glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
-}
-
-void VertexBuffer::Unbind() const
-{
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
-}
